tipcow-rewards/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.3.2
SQLAlchemy==1.1.15
Werkzeug==0.12.2
PyYAML